+++ R/dbi-driver.R
@@ -48,6 +48,10 @@ setMethod("show", "OdbcDriver",
 #'   not using UTF-8 you will need to set the encoding to get accurate
 #'   re-encoding. See [iconvlist()] for a complete list of available encodings
 #'   on your system. Note strings are always returned `UTF-8` encoded.
+#' @param name_encoding The text encoding for column names used on the
+#'   Database.  May be different than the `encoding` argument.  Defaults to
+#'   empty string which is equivalent to returning the column names without
+#'   performing any conversion.
 #' @param driver The ODBC driver name or a path to a driver. For currently
 #'   available options, see the `name` column of [odbcListDrivers()] output.
 #' @param server The server hostname. Some drivers use `Servername` as the name
@@ -164,6 +168,7 @@ setMethod("dbConnect", "OdbcDriver",
       timezone = "UTC",
       timezone_out = "UTC",
       encoding = "",
+      name_encoding = "",
       bigint = c("integer64", "integer", "numeric", "character"),
       timeout = 10,
       driver = NULL,
@@ -179,6 +184,7 @@ setMethod("dbConnect", "OdbcDriver",
     check_string(timezone)
     check_string(timezone_out)
     check_string(encoding)
+    check_string(name_encoding)
     arg_match(bigint)
     check_number_decimal(timeout, allow_na = TRUE)
     check_string(driver, allow_null = TRUE)
@@ -199,6 +205,7 @@ setMethod("dbConnect", "OdbcDriver",
       timezone = timezone,
       timezone_out = timezone_out,
       encoding = encoding,
+      name_encoding = name_encoding,
       bigint = bigint,
       timeout = timeout,
       driver = driver,